+++ lib/jekyll/rp_logs/rp_tags.rb
@@ -8,13 +8,6 @@ module Jekyll
       CHAR_FLAG = /^char:(?<char_name>.*)/
       META_TAGS = /(safe|questionable|explicit|canon|noncanon|complete|incomplete)/
 
-      TAG_NAME_MAP = {
-        "#"  => "sharp",
-        "/"  => "slash",
-        "\\" => "backslash",
-        " "  => "_"
-      }.freeze
-
       TYPE_CLASSES = {
         character: ["rp-tag-character"],
         meta: ["rp-tag-meta"],
@@ -90,24 +83,10 @@ module Jekyll
 
       private
 
-      # Map a tag to its directory name. Certain characters are escaped,
-      # using the TAG_NAME_MAP constant, above.
+      # Map a tag to its directory name. Unsafe characters are replaced with
+      # underscores. This restricts the dir name to safe characters in URLs.
       def name_to_dir(name)
-        s = ""
-        name.each_char do |c|
-          if c =~ /[-A-Za-z0-9_|\[\]]/
-            s += c
-          else
-            c2 = TAG_NAME_MAP[c]
-            unless c2
-              msg = "Bad character '#{c}' in tag '#{name}'"
-              puts("*** #{msg}")
-              raise Exception.new(msg)
-            end
-            s += "#{c2}"
-          end
-        end
-        s
+        name.gsub(/[^-A-Za-z0-9_|\[\]]/, "_")
       end
     end
   end