+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Upcoming
+### Changed
+- Certain characters (`#/\`) are replaced with `_` now instead of a word escape.
+
+### Fixed
+- Unsafe characters for URLs are replaced with `_` in tags now instead of throwing an exception. ([#48], [#49])
+
 ## [0.3.0] - 2015-11-06
 ### Added
 - Tags are displayed underneath the title of an RP on its page (requires theme update) ([#37])
@@ -86,3 +93,5 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 [#12]: https://github.com/xiagu/jekyll-rp_logs/issues/12
 [#36]: https://github.com/xiagu/jekyll-rp_logs/issues/36
 [#37]: https://github.com/xiagu/jekyll-rp_logs/issues/37
+[#48]: https://github.com/xiagu/jekyll-rp_logs/issues/48
+[#49]: https://github.com/xiagu/jekyll-rp_logs/issues/49